# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EduSearch API",
    "description": "Comprehensive Indian Education Data API — JEE/NEET cutoffs, NIRF rankings, college data, placements, and admission predictions.",
    "version": "1.0.0",
    "contact": {
      "name": "EduSearch API"
    }
  },
...
  "paths": {
...
    "/api/v1/colleges": {
      "get": {
        "summary": "Search and filter institutes",
        "tags": [
          "Colleges"
        ],
        "description": "Returns institutes matching the given filters, ordered by NIRF rank then name. Use it to build pickers and browse pages — for example `?type=IIT&state=Tamil Nadu` or `?nirf_rank_max=10`. The `id` on each record is what /api/v1/colleges/{id} and /api/v1/compare expect.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "IIT",
                "NIT",
                "IIIT",
                "GFTI",
                "Medical"
              ]
            },
            "in": "query",
            "name": "type",
            "required": false,
            "description": "Restrict to one institute category."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "state",
            "required": false,
            "description": "Exact state name, e.g. \"Maharashtra\"."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "city",
            "required": false,
            "description": "Exact city name, e.g. \"Mumbai\"."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false,
            "description": "Partial match against the full name or abbreviation."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "query",
            "name": "nirf_rank_min",
            "required": false,
            "description": "Only institutes ranked at or below this number (worse than or equal to)."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "nirf_rank_max",
            "required": false,
            "description": "Only institutes ranked at or above this number (better than or equal to)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "has_program",
            "required": false,
            "description": "Only institutes offering a program matching this text, e.g. \"Computer Science\"."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Maximum records to return (1-200)."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Records to skip, for paging through results."
          }
        ],
        "responses": {
          "200": {
            "description": "Matching institutes with pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Matching institutes with pagination metadata.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always true on a successful response."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Unique institute ID. Use this with /api/v1/compare and /api/v1/colleges/{id}."
                          },
                          "name": {
                            "type": "string",
                            "description": "Full official name, e.g. \"Indian Institute of Technology Bombay\"."
                          },
                          "short_name": {
                            "type": "string",
                            "nullable": true,
                            "description": "Common abbreviation, e.g. \"IIT Bombay\"."
                          },
                          "type": {
                            "type": "string",
                            "nullable": true,
                            "description": "Institute category: IIT, NIT, IIIT, GFTI or Medical."
                          },
                          "state": {
                            "type": "string",
                            "nullable": true,
                            "description": "State the institute is located in."
                          },
                          "city": {
                            "type": "string",
                            "nullable": true,
                            "description": "City the institute is located in."
                          },
                          "nirf_rank": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Most recent NIRF rank, or null if unranked."
                          },
                          "nirf_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Most recent NIRF score out of 100."
                          },
                          "website": {
                            "type": "string",
                            "nullable": true,
                            "description": "Official website URL."
                          },
                          "established_year": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Year the institute was founded."
                          }
                        }
                      },
                      "description": "The matching records."
                    },
                    "meta": {
                      "type": "object",
                      "description": "Pagination metadata.",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "description": "Total records matching the filters, ignoring limit and offset."
                        },
                        "limit": {
                          "type": "integer",
                          "description": "Maximum records returned in this response."
                        },
                        "offset": {
                          "type": "integer",
                          "description": "Number of records skipped."
                        },
                        "has_more": {
                          "type": "boolean",
                          "description": "True when more records remain beyond this page."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "One or more query parameters were invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "One or more query parameters were invalid.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/colleges/{id}/placements": {
      "get": {
        "summary": "Get placement records for an institute",
        "tags": [
          "Colleges"
        ],
        "description": "Returns placement figures — placement percentage and median, average and highest packages — for one institute, newest year first. Add `?year=` to pin a single season.",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "year",
            "required": false,
            "description": "Restrict to a single placement year, e.g. 2025."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Institute ID, as returned by /api/v1/colleges."
          }
        ],
        "responses": {
          "200": {
            "description": "Placement records for the institute.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Placement records for the institute.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always true on a successful response."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Unique placement record ID."
                          },
                          "institute_id": {
                            "type": "integer",
                            "description": "ID of the institute."
                          },
                          "year": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Placement season year."
                          },
                          "program_or_dept": {
                            "type": "string",
                            "nullable": true,
                            "description": "Program or department the figures cover."
                          },
                          "students_placed_pct": {
                            "type": "number",
                            "nullable": true,
                            "description": "Share of eligible students placed, as a percentage."
                          },
                          "median_salary": {
                            "type": "number",
                            "nullable": true,
                            "description": "Median annual package in INR."
                          },
                          "average_salary": {
                            "type": "number",
                            "nullable": true,
                            "description": "Average annual package in INR."
                          },
                          "highest_salary": {
                            "type": "number",
                            "nullable": true,
                            "description": "Highest annual package in INR."
                          },
                          "top_recruiters": {
                            "type": "string",
                            "nullable": true,
                            "description": "JSON-encoded array of recruiter names."
                          }
                        }
                      },
                      "description": "Placement records, newest year first."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The institute ID or year was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The institute ID or year was invalid.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
}